Allow query plan optimisation to reuse progress when statistics change (#6609)
## What is the goal of this PR?

We improve the perofrmance of the query planner on repeat invocations, even in the event of the changing traversal costs by saving and restarting the planner from the last valid solution.

## What are the changes implemented in this PR?

Previously, when the data statistics had changed enough that the planner was no longer considered up to date, one of two things would happen. If the solver had reached the optimal state and was released, a new solver was created and initialised with the same initialisation, essentially restarting the search from square one. Alternately, if the solver was still present, the objective function changes were not communicated to it (!), and if the constraints changed, it was potentially in an invalid state, i.e. the variable values did not satisfy the constraints, which effectively restarted the search as well.

The main change in this PR is communicating the changes to the solver directly from the planner's side under the assumption that the previously made optimisation progress would be useful wrt the changed model as well. Effectively, we reset the optimiser and initialise it to the previously found solution, adjusted to satisfy the updated constraints.

* `OptimiserVariable` no longer separates the initial value from the solution value. This allows the previously recorded solution value to be seamlessly reused as the new solver hint, when required.
* `GraphPlanner` informs the `Optimiser` directly whether its model needs updating. This removes the need for "dirty" flags, and ensures that the solver state is fully updated when the objective function coefficients and constraints change.