Enforce query optimisation time limit using background planning (#6616)
## What is the goal of this PR?

Strictly enforce a time limit for query planning to create more predictable query execution latencies. Query planning using the OR-Tools SAT solver may not respect our fine-granularity time limits (eg. 100ms), and may overrun up several multiples of this before stopping. We now plan in a background thread, and use the an updated plan if it is available within the time limit, and otherwise fall back to the previous query plan.

## What are the changes implemented in this PR?
* Query planning executes in a background thread using `CompletableFuture`
* Refactor `GraphPlanner.mayOptimise()` to allow first-time optimisation or re-optimisation to utilise the same `Future`
* All threads will now be allowed to wait up to the `timeLimit` for the optimisation to finish. Previously, if there was a re-optimisation running we actually always used the old plan!