Reasoner tracing: Fix initialization race & handle case when directory does not exist (#6615)
Reasoner tracing: Fix initialization race & handle case when directory already exists
## What is the goal of this PR?
Fix bugs occurring whe reasoner tracing is used in multiple transactions during the same session.

## What are the changes implemented in this PR?
The TraceWriter member used to be initialised lazily by the traceWriter method when first accessed. This has a concurrency bug which is triggered with multiple writers. Now it  is initialised in the constructor of Tracer.

Also, initialisation would fail when the output directory already exists. This is fixed with a check for directory existence.