Fix reasoner to cache plans for given bounds (#6621)
## What is the goal of this PR?
The optimal plan (resolvable ordering)  for a conjunction depends on which variables are bound. 
Currently, the reasoner controllers cache a single plan for a given conjunction, irrespective of the bounds - (though the plan function accepts the bound variable set as an argument)

## What are the changes implemented in this PR?
This PR updates conjunctionController to cache plans per bounds, using a map of bounds -> plan.