Fix storage iterator forwarding (#6623)
## What is the goal of this PR?

We fix the behaviour of our storage iterators when forwarding them to a target. Storage iterators are bound to a particular prefix internally. However forwarding them to a value before the prefix should still be a valid operation that returns the entire prefix range as before, rather than returning nothing.

## What are the changes implemented in this PR?

* Previously, forwarding a storage iterator to a value before the prefix would be allowed to proceed to performing a `seek()` operation on RocksDB
* However, this mean the next value retrieved from RocksDB wouldn't have the prefix that our storage iterator was bound to, and the storage iterator would immediately return with no answers
* Now, we do a no-op if the target is before the prefix that the iterator is bound to.