Unifying a concludable with a conclusion considers value constraints (#6628)
## What is the goal of this PR?
The current rule-validation incorrectly flags valid & useful scenarios as a negative cycle, such as in #6538 
These occur when the conclusion is a `has` with a constant attribute value, and the concludable has constraints with constant values. This PR incorporates evaluating the constraints in the concludable against the concluded value in such cases.  

## What are the changes implemented in this PR?
* Extends Concludable.Has.unificationSatisfiable to evaluate constant value constraints against constant-value conclusions 
* Implements  the constraint consistency check in pattern::ValueConstraint.
- Currently only supports the case where one constraint is the an exact equality. The other can be any constraint. (Thus, the implementation checks whether a value satisfies a constraint, rather than checking whether two constraints are compatible)
* Extends unification requirements to include other value constraints in addition to equality.

This would satisfy/ fix #6538