This commit contains 3 smaller ones: -------- Move logic caching responsibilities to LogicManager & LogicCache (#6632)
We try to bring the logic-package to reflect the pattern package (with the eventual, far-in-the-future goal of unifying the two). We also move the caching outside of the logical-objects (Rules & Resolvables) and make it the responsibility of the LogicManager and LogicCache, and introduce the logic-package equivalents of Disjunctions, Conjunctions and Negations to replace the sometimes-complicated interaction between patterns and resolvables. This simplification gives us the elements needed to create an abstract representation of the structure of the reasoner's 'theory'.

This hopefully allows us clean up the logic & reasoning code a bit before we think about unifying with the pattern package.

* Remove rule-graph related caching from the Concludable &  Rule classes.
* Introduce logic equivalents of Conjunctions, Disjunctions & Negations
* Move rule-graph related functionality from Rule to LogicManager, so the Rule instance can focus on itself.

---------
Refactor reasoner graph construction to use logic data-structures  (#6626)

We update the reasoner controllers to use the (new) logic conjunction & disjunctions in place of the patterns - enabling a cleaner mapping between the structures of the logical-theory (conjunctions, concludables etc) and the computation-graph.

The reasoner views Conjunctions as a set of Resolvables. Because conjunctions are concepts in the lower 'pattern' domain and Resolvables are concepts from the higher logic domain, conjunctions do not store this mapping.

Since  Resolvables nor Conjunctions have a reliable equality test, we rely on object identity and pass the objects themselves around when required. This is limiting in the reasoner - If we need the constituent concludables of a conjunction - we would need to pass them around with the conjunction.

* Refactor Controllers & Processors use ResolvableConjunction, ResolvableDisjunciton in place of corresponding patterns
------------

Refactor reasoner planner to be able to use statistics (#6620)

We move the reasoner-planner to an architecture which enables recursive computation of costs through triggered rules.

We also implement a version which emulates the old local heuristic planner by returning appropriate dummy costs. This can serve as a baseline to compare against.

A Plannable class maps the logical concepts to the planning domain - We consider the unit of planning to be a logical entity + the set of bound variables it receives. We capture this in a **Plannable** class.

We create an abstract **ReasonerPlanner** which serves as a coordinator the planning - allowing a Plannable to access the cost of the Plannables it recursively depends on. The actual planning algorithm is left to the subclasses to implement.

Finally, we implement a GreedyAnswerSizeSearch subclass which (in the next phases) will be developed into the simplest prototype plan search based on a cost estimated from data statistics. At this point, the cost is set to 1 for any Plannable.

Porting the old planning algorithm to the new architecture, we implement a new ReasonerPlanner subclass - **OldPlannerEmulator** which returns the same result as the old-planning algorithm would. This will serve as a sanity-check and a good base-line in future, as well as allowing us to re-use the existing integration tests to cover some of the new code-paths.