Move logic caching responsibilities to LogicManager & LogicCache (#6632)
## What is the goal of this PR?
We try to bring the logic-package to reflect the pattern package (with the eventual, far-in-the-future goal of unifying the two). We also move the caching outside of the logical-objects (Rules & Resolvables) and make it the responsibility of the LogicManager and LogicCache, and introduce the logic-package equivalents of Disjunctions, Conjunctions and Negations to replace the sometimes-complicated interaction between patterns and resolvables. This simplification gives us the elements needed to create an abstract representation of the structure of the reasoner's 'theory'.  

This hopefully allows us clean up the logic & reasoning code a bit before we think about unifying with the pattern package. 

## What are the changes implemented in this PR?
* Remove rule-graph related caching from the Concludable &  Rule classes.
* Introduce logic equivalents of Conjunctions, Disjunctions & Negations
* Move rule-graph related functionality from Rule to LogicManager, so the Rule instance can focus on itself.