Remove redundant commit rule validation (#6634)
## What is the goal of this PR?

Eliminate an expensive rule re-validation on schema transaction commit when the schema types have not changed. Rules are therefore only validated at define time, and re-validated at commit time if the types have changed.

If using rules that are costly to validate, the optimal way to define a schema is now to use one transaction to create all the Types, and a separate transaction to create all the Rules. 

## What are the changes implemented in this PR?

* Only re-validate rules at commit time if any schema types have been modified during the transaction

Fixes #6633