Refactor reasoner graph construction to use logic data-structures  (#6626)
## What is the goal of this PR?
We update the reasoner controllers to use the (new) logic conjunction & disjunctions in place of the patterns - enabling a cleaner mapping between the structures of the logical-theory (conjunctions, concludables etc) and the computation-graph.
 
The reasoner views Conjunctions as a set of Resolvables. Because conjunctions are concepts in the lower 'pattern' domain and Resolvables are concepts from the higher logic domain, conjunctions do not store this mapping. 

Since  Resolvables nor Conjunctions have a reliable equality test, we rely on object identity and pass the objects themselves around when required. This is limiting in the reasoner - If we need the constituent concludables of a conjunction - we would need to pass them around with the conjunction.

## What are the changes implemented in this PR?
* Refactor Controllers & Processors use ResolvableConjunction, ResolvableDisjunciton in place of corresponding patterns