Refactor reasoner planner to be able to use statistics (#6620)
## What is the goal of this PR?
We move the reasoner-planner to an architecture which enables recursive computation of costs through triggered rules.

We also implement a version which emulates the old local heuristic planner by returning appropriate dummy costs. This can serve as a baseline to compare against. 

## What are the changes implemented in this PR?
A Plannable class maps the logical concepts to the planning domain - We consider the unit of planning to be a logical entity + the set of bound variables it receives. We capture this in a **Plannable** class.

We create an abstract **ReasonerPlanner** which serves as a coordinator the planning - allowing a Plannable to access the cost of the Plannables it recursively depends on. The actual planning algorithm is left to the subclasses to implement.  

Finally, we implement a GreedyAnswerSizeSearch subclass which (in the next phases) will be developed into the simplest prototype plan search based on a cost estimated from data statistics. At this point, the cost is set to 1 for any Plannable.

Porting the old planning algorithm to the new architecture, we implement a new ReasonerPlanner subclass - **OldPlannerEmulator** which returns the same result as the old-planning algorithm would. This will serve as a sanity-check and a good base-line in future, as well as allowing us to re-use the existing integration tests to cover some of the new code-paths.