Optimise sort queries by utilising sorted data on disk (#6641)
## What is the goal of this PR?

TypeDB can leverage the sorted nature of the data storage engine for non-string attributes, to implement optimised sort queries. Previously, TypeDB would inefficiently gather all answers to a query, sort them, and then output them.

TypeDB also now supports sorting different each variable in their own direction (eg. ascending or descending).

## What are the changes implemented in this PR?

The main change is that TypeDB pins the sort variables at the start of the query execution
  * This guarantees that the sort variables are produced in order!
  * The rest of the query is traversed from the fixed start points, guaranteeing we find all answers in the user's sort order
  * the query planner implements this with new ordering constraints, and an updated plan initialiser  
  * we introduce a validation before any of query executes to ensure that all concepts retrieved within the sorted variable are comparable - eg. if `$a` can be `datetime` and `long`, these are not comparable, so the query is rejected.

We also push the traversal modifiers (`filter` and `sort`) into the query planner and traversal execution, and wrap them in new server-side constructs nested in `//traversal/common:Modifiers`. This enables the query planner to be aware fo the sorting requirements, and in the future of the filtering requirements as well. In addition, the traversal can then use the right ascending or descending sort order for the sorted vertices.

`ConceptMap` has a new comparable subtype called `ConceptMap.Sortable`, which contains a comparator for sorting the concept maps according to some of the variables in the concept map. This allows us to use our standard `SortedIterator` implementations to handle ordered concept map streams.
* as a result, we also use `? extends ConceptMap` wherever the `Matcher.execute()` returns to handle the possibility of getting sortable concept maps.
* we also move the comparators used to sort concept map lists into `ConceptMap.Comparator`

Other changes
* update `FunctionalIterator`s and `SortedIterator`s to allow merge sorting and mapping more freely, as well as renaming to disambiguate methods overloaded purely generic parameters, which Java can't otherwise distinguish
* remove redundant `Context` arguments throughout the querying APIs, such as in `Reasoner.java`
* low-level APIs in the `//graph` package have optional arguments to enforce an ordering of the returned iterator of vertices. For example, `Forwadable<ThingVertx, ORDER> thingGraph.getReadable(TypeVertex type, ORDER order)` now has an order argument and parameter