Reasoner planner treats IID and value-equality constraints as bounds (#6650)
## What is the goal of this PR?
The reasoner planner must treat constraints binding a variable to a specific value (or IID) similar to how it treats the variable being bound to a specific thing.
 
When planning the order of resolvables constituting a conjunction, the current reasoner-planner prioritises a connected resolvable over a disconnected one.
A variable which is bound to an IID (or a specific value) has a similarly strong restricting effect. With this PR, the reasoner planner treats these variables as being bound (and hence, all resolvables featuring them as connected) AFTER it runs out of connected resolvables. (Note: for the top-level query, this is always the case).  

## What are the changes implemented in this PR?
- Reasoner planner: When all resolvables left to schedule are disconnected, the boundVariables are extended to include the variables bound to an IID or a specific value.