Answer count model for reasoner planner (#6640)
## What is the goal of this PR?

We introdue statistics-based answer count estimation which will be the basis of the cost of reasoner plans.

For a given conjunction, We aim to estimate:
* The number of answers for a given (sub)set of variables
* when considering only a subset of resolvables of the conjunction 
We believe this is sufficient for our formulation of the reasoner planner's cost function. 

## What are the changes implemented in this PR?
* **reasoner.planner.AnswerCountEstimator.ConjunctionModel:**
  * Given a conjunction, builds a model and answers queries to estimate answers for 
    * The number of answers to a subset of variables
    * Considering only a subset of the specified resolvables in the conjunction. 
* **reasoner.planner.AnswerCountEstimator.LocalModel:**
  * The units of estimation. The ConjunctionModel derives its estimates from the LocalModels 
* **reasoner.planner.AnswerCountEstimator:**
  * Registry of ConjunctionModels
  * Coordinates between them while inter-dependent models are being built
* **reasoner.planner.AnswerCountEstimatorTest:**
  * TestCases for concludables & conjunctions verifying the estimates are as expected.

## How it works
### The conjunction model
Rules are inherently recursive, and a planner will ask for answerEstimates for various variable and resolvable subsets multiple times - To prevent expensive computation, we build conjunction models which can answer these queries without recursing into triggered rules. 

We describe the model by starting at a naive estimate and incrementally refining the model. We use as running example:
```
$p isa person; $c isa person; $d isa dog;
$_pc (parent: $p, child: $c) isa parentship; $_pd (owner: $c, pet: $d) isa pet-ownership ;
```
We'll query the variables {$p,$c,$d} (and use $_pc, and $_pd for easy referencing )   

#### Variable type-information only
For a given conjunction, the number of answers to the variables is the number of values each variable can take-up. Take the product across variables, of the number of things in the type-graph matching the variable set of types.

`estimate({$p, $c, $d))  = thingsMatchingTypes({person}) * thingsMatchingTypes({person}) * thingsMatchingTypes({dog})`

#### Resolvables constrain variables
Notice the resolvables restrict the number of values a variable can take - $p is restricted to the number of role-players of parent in a parentship.

```
estimate({$p, $c, $d))  =
  min(thingsMatchingType({person}), rolePlayers(parentship:parent)) * 
  min(thingsMatchingType({person}, rolePlayers(parentship:child), rolePlayers(pet-ownership:owner)) *
  min(thingsMatchingType({dog}), rolePlayers(pet-ownership:pet) 
```

#### Factorising estimates over the resolvables:
It's become obvious above that the rolePlayer constraints (and hasEdge constraints, etc) are enforced only when a resolvable is in the evaluated set.
The conjunction model uses this fact by constructing a 'local model' for each resolvable (using numbers from stats and an approximate recursive answer count (see acyclic-models)). The answer count to a variable given a set of resolvables to evaluate is the minimum, across all local models, of the answers to the variable returned by each model.

#### Co-constrained variables
$p and $c  (and $c,$d) are co-constrained by being in the same relation. 
Roughly, moving our estimate from factorising over variables to factorising over resolvables looks like this:
```
estimate = min {
   answers($_pc, {$p,$c}) * answers($_pd, {$d}),
   answers($_pc, {$p}) * answers($_pd, {$c, $d})
}
```
We compute this (approximately) by doing a greedy cover - Find the set of resolvables which cover all the queried variables but has the minimal answer count.   

### Building local models and handling cycles
Building local models for concludables becomes complicated when they trigger a cycle - since the answer-count recursively depends on itself.
Building the local model for a concludable which is not in a cyclic path is simpler - we can use the conjunction models of the conditions of the triggered rules to estimate the inferred count. 

We figured answer-counts involving cycles are hard, and we approximate it by breaking the cycle: 
When constructing the model for a cyclic concludable, we only consider the acyclic concludables in the triggered rules.

This leads us to a three-phase process for building a model for a conjunction

1. Register conjunction & its dependencies, identify cyclic concludables
2. Construct acyclic models for all of them
3. Construct cyclic models based on the acyclic model estimates