Linearise query traversal (#6660)
## What is the goal of this PR?

We ensure that the traversal of a query procedure is performed with as little "jumping" between unrelated concepts as possible.

## What are the changes implemented in this PR?

When a graph iterator fetches the candidates for ea, it does so in the order specified in the procedure. Because the graph planner only considers the direction of the edges and not the exact ordering of the vertices, the resulting procedure may look as follows:

```
   0
 /   \
1     2
|     |
4     3
```

In this example, if the vertex no. 4 does not match any things in the database, the iterator would backtrack back to vertex no. 1 and revisit vertices no. 2 & 3, duplicating work unnecessarily.

Through a simple depth-first topological sort, we ensure that the traversal follows edges whenever possible, eliminating this problem.