Improve concurrency of TypeDB executors (#6667)
## What is the goal of this PR?

We fix a bug in the allocation of jobs onto executor pools, which was
limiting high-CPU environments from achieving maximum concurrency. In
some workloads this change increased CPU utilisation from 40% to 65%.

## What are the changes implemented in this PR?

* Allocation of a job onto a `ParallelThreadPoolExecutor` considers not
only the job queue size currently, but also the number of actively
executing jobs. This means that we don't end up queuing a job behind an
active job when there is another executor that is fully un-used