Improve write performance by removing persisted statistics version (#6692)
## What is the goal of this PR?

We fix a performance bottlenecks during bulk-writes: the continuous persistence and reading of the statistics version number. We completely remove the on-disk statistics version, and only maintain an in-memory statistics version number. This improves pure write throughput by up to 2x.

## What are the changes implemented in this PR?

#### Problem

Each write transaction uses RocksDB `merge` to increment the statistics snapshot number. However, each transaction also reads the snapshot version number (note that each transaction is snapshotted, so will see its own statistics snapshot number). 

However, over 70% of a transaction time was spent reading the version number on transaction open when a huge number of writes are happening to the DB (normally when doing straight writes without reads). We suspect this is because RocksDB has to deal with both snapshotting & flatten all atomic increment-and-set on a key-value pair in the LSM tree each time the statistics snapshot read occurs.

#### Solution

We realise that the statistics verison number of is used for updating query plans when the statistics have changed. However, the query plans are shared across all transactions, and not bound to a particular transaction or data snapshot. So, it makes sense to replan whenever the database statistics version has changed -- we don't need to keep the snapshot version bound to a transaction, but the database.

Additionally, query plans are discarded on shutdown - so the statistics snapshot number does not have to be persisted. 

Instead, we no keep an in-memory `AtomicLong` that represents the statistics version of the database. A write transaction that commits any statistics to disk will induce an increment the in-memory `statisticsVersion`. Query planning in turn reads the database statistics version instead of a transaction statistics version.