Type inference differentiates illegal and queries with no answer (#6693)
## What is the goal of this PR?

We update the usage of type inference to differentiate between an
impossible-to-satisfy query (eg. using relations between types that are
not playing the right roles, mismatched value types, etc.) and a query
that will not have any answers but is semantically sound - such as
querying for instances of an abstract type that has no concrete
subtypes.

## What are the changes implemented in this PR?

* Type inference sets two booleans on `Conjunction`: `isCoherent` is set
to `false` when the query represents a semantic schema violation.
`isAnswerable` is set to `false` when the conjunction will never create
any answers in the current schema.
* Short-circuit the execution of a query that does not have any
answerable conjunctions and return an empty iterator

To implement these changes, we run type inference twice.
1. run type inference, including abstract types in the type inference
output
2. If no types are found, then the query is illegal and we throw an
exception
3. Run type inference disallowing abstract types. If any variables do
not have any types, we set the query ans unanswerable

This restores the built-in assumption implementation of traversal,
planning, reasoner, rule validation etc. that only queries where each
variable contains a non-empty set of concrete types on each variable
shall be processed.