Relaxing RelationType commit validation
Now that we validate when a role type overrides a valid super type that
it inherits from its relation type, we are able to relax the commit
validation of RelationType. The validation now filters out root role
type when checking for abstract role types that are related in the
relation type. The only scenario where you can inherit a root role type
while not having empty role types, is when the the related role type
doesn't override an inherited role type in the relation, which will be
caught by the role type commit validation.