Simplify isolation manager (#6704)
## What is the goal of this PR?

We discover that a ConcurrentSkipListMap does not implement
locking/atomic `compute` operations. This behaviour was relied on to
implement Isolation Manager's cleanup of transaction state, and errors
could manifest as concurrent modification exceptions. To fix this, we
refactor the IsolationManager with a simpler state model to optimally
allow both isolation validation and deletion.


## What are the changes implemented in this PR?

* `IsolationManager` cleanup of committed transactions no longer relies
on a `timeline`, but just the current set of transactions and the state
they are in. We also delete the single concurrent hash map of
`CommitState`
* Introduce three maps: `uncommitted`, `committing`, and `committed`.
These allow us to optimise the single-threaded lock duration in the
isolation manager, and implement a simplified cleanup mechanism.