Optimise traversal over non-retrieved variables (#6707)
## What is the goal of this PR?

Query plans which retrieve a tail sequence of vertices to visit, which are not in the user's query filter, can short-circuit their backtracking after a full answer is found. This means only permutations of vertices up to the last retrieved vertex are generated. This can dramatically speed up large queries in which only a few variables are retrieved. This also noticably speeds up rule validation, where only a few variables of the `when` need to be retrieved.

## What are the changes implemented in this PR?

To illustrate the implementation: given ordered traversal variables [A, B, C, D, E], of which only [A, C] are retrieved by the query filter, after finding a successful permutation for all vertices D and E can immediately backtrack up to C, since they are not in the query filter.

Code changes:
* Implement optimised backtracking through un-retrieved vertices in the traversal engine